Divide SdmGoal into SdmGoalEvent and SdmGoalMessage
The message is what we store; the event is what we receive from subscriptions or queries
