+++ src/api-helper/goal/sdmGoal.ts
@@ -16,13 +16,16 @@
 
 import { sprintf } from "sprintf-js";
 import {
-    SdmGoal,
     SdmGoalKey,
-} from "../../api/goal/SdmGoal";
+} from "../../api/goal/SdmGoalMessage";
+import { SdmGoalEvent } from "../../api/goal/SdmGoalEvent";
 
-export function isSdmGoal(a: object): a is SdmGoal {
-    const maybe = a as SdmGoal;
-    return !!maybe.goalSet && !!maybe.uniqueName && !!maybe.repo;
+/**
+ * @deprecated because I don't think we need this
+ */
+export function isSdmGoal(a: object): a is SdmGoalEvent {
+    const maybe = a as SdmGoalEvent;
+    return !!maybe.goalSet && !!maybe.uniqueName && !!maybe.name && !maybe.branch;
 }
 
 export function mapKeyToGoal<T extends SdmGoalKey>(goals: T[]): (k: SdmGoalKey) => T {
