+++ src/graphql/fragment/SdmGoalFields.graphql
@@ -4,6 +4,9 @@ fragment SdmGoalFields on SdmGoal {
   name
   sha
   branch
+  push {
+    ...PushFields
+  }
   fulfillment {
     method
     name
